Reduce occurence of Deck keyboard crash
Background: https://github.com/Plagman/gamescope/issues/724

This doesn't eliminate it, but should reduce the frequency of hitting
that assert. It's still undefined behaviour if another surface
happens to get this same ID, but this is still an improvement, although
it required expoing some of the internals of wlroots to do so.